+++ lib/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+
+- `include` config option for overriding `ignore`
 
 ## [0.1.0] - 2024-04-28